added a new parameter in the post request and it is now checked if the Stopped file is of the displayed format

+++ website/files/deleteAllStoppedFiles.php
@@ -6,13 +6,14 @@ require_once '../controllers/SessionController.php';
 use classes\FilesController;
 use classes\SessionController;
 
-if (isset($_POST) && isset($_POST['token'])) {
+if (isset($_POST) && isset($_POST['token']) && isset($_POST['filetype'])) {
+    $filetype = "." . $_POST['filetype'];
     $session = new SessionController($_POST['token']);
     session_start();
     $files = new FilesController($session->session_id);
 
     for ($fileNb = 0; $fileNb < $files->getFileListSize(); $fileNb++) {
-        if ($files->getFileStatus($fileNb) === "Stopped") {
+        if ($files->getFileStatus($fileNb) === "Stopped" && strpos($files->getFileName($fileNb), $filetype) !== false) {
             unlink($files->getFileRelativePath($fileNb));
         }
     }
